attempt at scanning for minimum
This doesn't appear to be working as intended. The idea was to make an array of the data at the end of the paths and scan through each to find the data points with the same time value allowing the calculation of the width at particular times
My next idea is to just take periodic slices based on time.